wallet_api/core/views.py
from rest_framework.viewsets import ModelViewSet
from rest_framework_json_api.pagination import JsonApiPageNumberPagination
from .models import Wallet, Transaction
from .serializers import WalletSerializer, TransactionSerializer

class WalletViewSet(ModelViewSet):
    queryset = Wallet.objects.all()
    serializer_class = WalletSerializer
    pagination_class = JsonApiPageNumberPagination

class TransactionViewSet(ModelViewSet):
    queryset = Transaction.objects.all()
    serializer_class = TransactionSerializer
    pagination_class = JsonApiPageNumberPagination
